Develop finish sprint14

+++ routes/cards.js
@@ -15,9 +15,24 @@ router.get('/', (req, res) => {
 });
 
 router.delete('/:id', (req, res) => {
-  Card.findByIdAndRemove(req.params.id)
-    .then((card) => res.send({ data: card }))
-    .catch(() => res.status(404).send(errRoute));
+  Card
+    .findById({ _id: req.params.id }, (err, data) => {
+      if (!data) {
+        return res
+          .status(404).send(errRoute)
+          .end();
+      }
+      if (!(data.owner.toString() === req.user.toString())) {
+        return res
+          .status(401)
+          .send({ message: 'Нет прав на удаление карточки' })
+          .end();
+      }
+      Card.findByIdAndRemove(req.params.id)
+        .then((card) => res.send({ data: card }))
+        .catch(() => res.status(404).send(errRoute));
+      return undefined;
+    });
 });
 
 router.post('/', createCard);
